requirements.txt
fastapi
sqlmodel
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]